aaa: Changing MaxLDAPQueryPartsNumber to long
Changed MaxLDAPQueryPartsNumber to be long to conform to
extensions-api.
In addition added support to Config for Long

topic: AAA

+++ backend/manager/modules/common/src/main/java/org/ovirt/engine/core/common/config/ConfigValues.java
@@ -728,7 +728,7 @@ public enum ConfigValues {
     @DefaultValueAttribute("1000")
     LdapQueryPageSize,
 
-    @TypeConverterAttribute(Integer.class)
+    @TypeConverterAttribute(Long.class)
     @DefaultValueAttribute("100")
     MaxLDAPQueryPartsNumber,
